third_party/blink/web_tests/wpt_internal/prerender/navigator-subapp.https.html

<!DOCTYPE html>
<title>navigator.subApp API test</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
promise_test(async () => {
  const bc = new BroadcastChannel('prerender-channel');

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  startPrerendering(`resources/subapp.html`);
  const result = await gotMessage;
  assert_equals(result, 'InvalidStateError');
}, 'prerendering pages should not be able to access subapp API.');
</script>
</body>
